Created the main file and placed the "player" character
There is now the proper software architecture to easily allow for more sprites to be implemented. Whether that's adding to an existing sprite sheet or adding new sprite sheets entirely.

Currently, there is only one sprite sheet containing one sprite for the player. The program spawns the player on screen and is allowed to move around.