Added custom math file and Player::shoot() function
The custom math file is for any useful math functions that are not already implemented in cmath. So far it only include gcd(), which finds the greatest common divisor between two numbers.

In addition, the shoot function has now been implemented. Currently, it only draws a single line from the center of the window to the edge in the direction of the mouse relative to the center of the screen.