Modified Character::collide() and move() functions, as well as Map::setCentrePlayer()
collide() and move() have been modified such that when an entity collides with another entity, they are still able to "rub" against the colliding hitbox, rather than simply freezing until they move int a completely different direction.

SetCentrePlayer() has been modified such that if the player is near the boundaries of the map, the map will not render such that it is possible to see outside its boundaries.