Implemented new structures to allow for the Player to carry a melee cone.
The plan is to use this cone to help detect entities in front of player and determine who the player damages when they use their melee weapon. Currently, it draws the melee cone of the "fists" weapon and draws it on the screen. It successfully updates it's position and orientation corresponding to the status of the player and mouse.