Character's do not bypass each other's hitboxes... to an extent.
If the character is moving in a straight line, it will not bypass the boundaries of the hitbox. However, it seems that when the character hitbox is in contact with hitboxes on different sides and they don't move in a straight line, they will bypass the hitbox boundaries. Currently working on a fix.