Enemies now more properly do no bypass each other's hitboxes... kinda.
Now they won't actively be inside each other's hitboxes. However, to achieve this, it was neccesary to implement an algorithm in which it detects whether each character's hitbox was inside another characters hitbox and then place them just outide of closest hitbox side. As a result, it is very common for characters to simply teleport of a bunch of them are grouped together.

Working on a better fix.